move status bar stuff to a different file
